Refactored package from entity to models for best practices
