+++ src/main/java/com/blackjack/models/Card.java
@@ -1,4 +1,4 @@
-package com.blackjack.entity;
+package com.blackjack.models;
 
 import com.blackjack.enums.Rank;
 import com.blackjack.enums.Suit;
