+++ src/test/java/com/blackjack/test/CardTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 
 import com.blackjack.enums.Rank;
 import com.blackjack.enums.Suit;
-import com.blackjack.entity.Card;
+import com.blackjack.models.Card;
 class CardTest {
 
 	@Test
