+++ src/test/java/com/blackjack/test/HandTest.java
@@ -4,10 +4,11 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import com.blackjack.models.Hand;
+
 import java.util.List;
 import java.util.ArrayList;
-
-import com.blackjack.entity.Hand;
 
 class HandTest {
 	private Hand hand;
